Specify bin width rather than number of bins
Compel the user to specify the width of the time and channel bins, rather
than the number of bins.  As the number of times present in a chunk of
Measurement Set rows is opaque in the general case, rather ask the user to
specify the number of samples that should be averaged into a single bin.