Integer factorisations of input channels
It's a good assumption that input channel widths are equal and we want
to maintain this in our output channelisations. For this to hold,
the output number of channels must be an integer factor
https://en.wikipedia.org/wiki/Integer_factorization
of the original number of channels, as this will divide
the original number of channels exactly.