Round up number of channels
This was confusing because previously, rounding down to select a smaller
channel width was used as the safe option to prevent decorrelation in
frequency. However, number of channels is the reciprocal of channel
width.

This therefore translates to selecting the safer option when choosing
the degree of channel averaging.  Put another way, our desired averaging
factors lies between two SPW bins.  Choose the one which doesn't
decorrelate.